Change command to run dev server
This was done for consistency reasons, everything can be ran with 'npm
run ...' instead of the dev serving being able to be ran with 'npm
start'. Not a huge deal but it's easy to forget the 'run' for build and
deploy considering how often we don't use the run keyword.